divides analyst helper folder based on what kind of object is the focus of the method
